ROE-2073 Updating DtoModelChangeTest to use custom mongo converters
