+++ src/test/java/uk/gov/companieshouse/overseasentitiesapi/configuration/MongoConfig.java
@@ -1,0 +1,9 @@
+package uk.gov.companieshouse.overseasentitiesapi.configuration;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class MongoConfig {
+
+    // Leave blank so that the tests pick up this config and not the real one.
+}
